+++ Makefile
@@ -228,10 +228,6 @@ RUSTFLAGS += --cfg=baseline_asterinas
 CLIPPY_COMMON_ARGS += --cfg=baseline_asterinas -A unused-imports -A dead-code -A unfulfilled-lint-expectations
 endif
 
-ifeq ($(CAPTURE_DATA), 1)
-RUSTFLAGS += --cfg=capture_data
-endif
-
 # To test the linux-efi-handover64 boot protocol, we need to use Debian's
 # GRUB release, which is installed in /usr/bin in our Docker image.
 ifeq ($(BOOT_PROTOCOL), linux-efi-handover64)
