+++ ostd/src/orpc/sync/mod.rs
@@ -5,7 +5,10 @@
 
 use alloc::vec::Vec;
 
+mod timeout_blocker;
+
 pub use orpc_macros::select;
+pub use timeout_blocker::TimeoutBlocker;
 
 use crate::{
     orpc::framework::CurrentServer,
